NEAT wumpus world unit test wip

+++ Galapagos/API/IPopulationMetadata.cs
@@ -21,7 +21,7 @@ namespace Galapagos.API
         /// <summary>
         /// Gets or sets the distance threshold for niches. A distance threshold of 0 disable niches.
         /// </summary>
-        uint DistanceThreshold { get; set; }
+        double DistanceThreshold { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating if cooperative coevolution should be used.
